apps/ai-pipeline/pyproject.toml
[project]
name = "ai-pipeline"
version = "0.1.0"
description = "AI pipeline for tree observation analysis — species ID, health assessment, measurements"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27,<1",
    "bullmq>=2.19,<3",
    "redis>=5.0,<7",
    "asyncpg>=0.29,<1",
    "minio>=7.2,<8",
    "Pillow>=10,<11",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5,<6",
    "respx>=0.21,<1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
